feat: implement feature to add  angle and speed to dataframe

+++ Computer Vision/DatasetEditor/DatasetEditor/model/DatasetImageLabel.cs
@@ -2,8 +2,8 @@ namespace DatasetEditor.model;
 
 public class DatasetImageLabel
 {
-    public int angle;
-    public int speed;
+    public double angle;  // it's an enum number (-3, -2, -1, 0, 1, 2, 3)
+    public double speed;  // it's an enum number (1, 2, 3, 4)
 
     public DatasetImageLabel(int angle, int speed) {
         this.angle = angle;
